+++ tests/got/test_default_reply_headers.js
@@ -25,7 +25,9 @@ describe('`defaultReplyHeaders()`', () => {
       'X-Powered-By',
       'Meeee',
       'X-Another-Header',
-      ['foo', 'bar'],
+      'foo',
+      'X-Another-Header',
+      'bar',
     ])
   })
 
@@ -34,10 +36,9 @@ describe('`defaultReplyHeaders()`', () => {
       'X-Powered-By',
       'Meeee',
       'X-Another-Header',
-      ['foo', 'bar'],
     ]
     nock('http://example.test')
-      .defaultReplyHeaders(defaultHeaders)
+      .defaultReplyHeaders([...defaultHeaders, ['foo', 'bar']])
       .get('/')
       .reply(200, '')
 
@@ -48,7 +49,7 @@ describe('`defaultReplyHeaders()`', () => {
       'x-another-header': 'foo, bar',
     })
 
-    expect(rawHeaders).to.deep.equal(defaultHeaders)
+    expect(rawHeaders).to.deep.equal([...defaultHeaders, 'foo', 'X-Another-Header', 'bar'])
   })
 
   it('default reply headers can be provided as a Map', async () => {
@@ -72,7 +73,9 @@ describe('`defaultReplyHeaders()`', () => {
       'X-Powered-By',
       'Meeee',
       'X-Another-Header',
-      ['foo', 'bar'],
+      'foo',
+      'X-Another-Header',
+      'bar',
     ])
   })