+++ types/index.d.ts
@@ -256,13 +256,13 @@ declare namespace nock {
     (
       fixtureName: string,
       options: BackOptions,
-      nockedFn: (nockDone: () => void) => void,
+      nockedFn: (nockDone: () => Promise<void>) => void,
     ): void
     (
       fixtureName: string,
       options?: BackOptions,
     ): Promise<{
-      nockDone: () => void
+      nockDone: () => Promise<void>
       context: BackContext
     }>
   }